Api integration

+++ Assets/Scripts/Screens/MainScreen/IMainScreenView.cs
@@ -17,7 +17,7 @@ namespace Screens.MainScreen
         public event Action RestartButtonClicked;
         public event Action<Vector2> EmptyScreenClicked;
 
-        void SetNextGiftTime(float timeToNextGift);
+        void SetNextGiftTime(int timeToNextGift);
         void SetIsMapActive(bool isMapActive);
         void SetCoins(int coins);
         void SetupActiveZone([CanBeNull] string zoneName);
